feat: step2 코드리뷰 반영 - datatest 추가
