+++ build.gradle.kts
@@ -1,3 +1,5 @@
+
+
 plugins {
     kotlin("jvm") version "1.8.10"
     id("org.jlleitschuh.gradle.ktlint") version "10.3.0"
@@ -14,6 +16,7 @@ dependencies {
     testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
     testImplementation("org.assertj", "assertj-core", "3.22.0")
     testImplementation("io.kotest", "kotest-runner-junit5", "5.5.3")
+    testImplementation("io.kotest", "kotest-framework-datatest", "5.5.3")
 }
 
 tasks {
